Remove stalls when navigating large projects
When clicking on a folder in the solution explorer that contained
many files (e.g. 5000), P4VS would hang due to the sheer number
of calls to File.Exists() in LocalPreferenceCache.RefreshAppData().

Checking if the file exists is unnecessary since File.GetLastWriteTime()
already returns Jan 1, 1601 when the file is missing. This value will
be newer than DateTime.MinValue, causing one initial load attempt
that will internally discover that the file does not exist.